Resolved all lint issues & renamed file & moved it where necessary
